% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_colors.R
\name{setup_colors}
\alias{setup_colors}
\title{Setup colors for xaringan document}
\usage{
setup_colors(...)
}
\arguments{
\item{...}{named parameters of colors to use.}
}
\value{
code to clipboard
}
\description{
Setup colors for xaringan document
}
\examples{
\dontrun{
setup_colors(red = "red", blue = "blue", green = "green")
}
}
